# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ademir API",
    "description": "API usada pelo novo frontend (React/Vite). Endpoints sob o prefixo /api.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:3101/api" }
  ],
...
  "paths": {
...
    "/estoque": {
      "get": { "tags": ["estoque"], "summary": "Obter estoque", "responses": { "200": { "description": "Estoque" } } },
      "post": {
        "tags": ["estoque"],
        "summary": "Salvar estoque",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object" } } } },
        "responses": { "200": { "description": "Estoque salvo" } }
      }
    },
    "/estoque/historico": {
      "get": { "tags": ["estoque"], "summary": "Histórico de estoque", "responses": { "200": { "description": "Histórico" } } },
      "post": {
        "tags": ["estoque"],
        "summary": "Salvar histórico",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "array" } } } },
        "responses": { "200": { "description": "Histórico salvo" } }
      }
    },
    "/estoque/verificar/{produtoNome}": {
      "get": {
        "tags": ["estoque"],
        "summary": "Verificar disponibilidade",
        "parameters": [{ "in": "path", "name": "produtoNome", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Disponibilidade" } }
      }
    },
    "/estoque/reduzir": {
      "post": {
        "tags": ["estoque"],
        "summary": "Dar saída em estoque",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "produtoNome": { "type": "string" }, "quantidade": { "type": "number" } } } } } },
        "responses": { "200": { "description": "Saída registrada" }, "400": { "description": "Estoque insuficiente" }, "404": { "description": "Produto não encontrado" } }
      }
    },
...
    "/controle/relatorio/{tipo}": {
      "get": {
        "tags": ["controle"],
        "summary": "Relatórios (cartoes, vendas, estoque, eventos)",
        "parameters": [{ "in": "path", "name": "tipo", "required": true, "schema": { "type": "string", "enum": ["cartoes", "vendas", "estoque", "eventos"] } }],
        "responses": { "200": { "description": "Relatório" }, "400": { "description": "Tipo inválido" } }
      }
    }
  }
}